Don't specify version when using version-file input (#10)
When using the `version-file` input, the version of the package to be
installed is derived from a `go.mod` file. Currently, the version is
taken from said file and then installed as normal.

However, using the `go.mod` file also means the transitive dependencies
are expected to come from the same file. When specifying a specific
version, the transitive dependencies come from the package's `go.mod`
file which may be older than the local `go.mod` file. This can cause
subtle changes, particularly when improvements have been made to AST
representations.

To fix, the version is dropped from the final `go install` command, and
support for executing from a different directory is added.

Fixes #8.